+++ query-engine/js-connectors/smoke-test-js/prisma/schema.prisma
@@ -4,9 +4,10 @@ generator client {
 }
 
 datasource db {
-  provider          = "@prisma/neon"
-  url               = env("JS_NEON_DATABASE_URL")
-  shadowDatabaseUrl = env("JS_NEON_SHADOW_DATABASE_URL")
+  // Once ran `prisma migrate reset` as per instructions in README.md replace `mysql` with `@prisma/planetscale` in the line below
+  provider          = "@prisma/planetscale"
+  url               = env("JS_PLANETSCALE_DATABASE_URL")
+  shadowDatabaseUrl = env("JS_PLANETSCALE_SHADOW_DATABASE_URL")
 }
 
 model some_users {
@@ -17,20 +18,26 @@ model some_users {
 
 model type_test {
   id                    Int                         @id @default(autoincrement())
+  tinyint_column        Int                         @db.TinyInt
+  tinyint_column_null   Int?                        @db.TinyInt
   smallint_column       Int                         @db.SmallInt
   smallint_column_null  Int?                        @db.SmallInt
+  mediumint_column      Int                         @db.MediumInt
+  mediumint_column_null Int?                        @db.MediumInt
   int_column            Int
   int_column_null       Int?
   bigint_column         BigInt
   bigint_column_null    BigInt?
-  float_column          Float                       @db.Real
-  float_column_null     Float?                      @db.Real
+  float_column          Float                       @db.Float
+  float_column_null     Float?                      @db.Float
   double_column         Float
   double_column_null    Float?
   decimal_column        Decimal                     @db.Decimal(10, 2)
   decimal_column_null   Decimal?                    @db.Decimal(10, 2)
   boolean_column        Boolean
   boolean_column_null   Boolean?
+  bit_column            Boolean                     @db.Bit(1)
+  bit_column_null       Boolean?                    @db.Bit(1)
   char_column           String                      @db.Char(10)
   char_column_null      String?                     @db.Char(10)
   varchar_column        String                      @db.VarChar(255)
@@ -41,14 +48,24 @@ model type_test {
   date_column_null      DateTime?                   @db.Date
   time_column           DateTime                    @db.Time(0)
   time_column_null      DateTime?                   @db.Time(0)
-  datetime_column       DateTime
-  datetime_column_null  DateTime?
+  year_column           Int                         @db.Year
+  year_column_null      Int?                        @db.Year
+  datetime_column       DateTime                    @db.DateTime(0)
+  datetime_column_null  DateTime?                   @db.DateTime(0)
   timestamp_column      DateTime                    @db.Timestamp(0)
   timestamp_column_null DateTime?                   @db.Timestamp(0)
   json_column           Json
   json_column_null      Json?
   enum_column           type_test_enum_column
   enum_column_null      type_test_enum_column_null?
+  binary_column         Bytes                       @db.Binary(64)
+  binary_column_null    Bytes?                      @db.Binary(64)
+  varbinary_column      Bytes                       @db.VarBinary(128)
+  varbinary_column_null Bytes?                      @db.VarBinary(128)
+  blob_column           Bytes                       @db.Blob
+  blob_null             Bytes?                      @db.Blob
+  set_column            String
+  set_column_null       String?
 }
 
 enum type_test_enum_column {
