+++ query-engine/js-connectors/smoke-test-js/src/neon.ts
@@ -1,8 +1,8 @@
 
 import { setImmediate, setTimeout } from 'node:timers/promises'
 
-import { binder } from './driver/util.js'
-import { createNeonConnector } from './driver/neon.js'
+import { binder } from './connector/util.js'
+import { createNeonConnector } from './connector/neon.js'
 import { initQueryEngine } from './util.js'
 
 async function main() {
@@ -14,18 +14,19 @@ async function main() {
   })
 
   // `binder` is required to preserve the `this` context to the group of functions passed to libquery.
-  const driver = binder(db)
+  const conn = binder(db)
 
   // wait for the database pool to be initialized
   await setImmediate(0)
 
-  const engine = initQueryEngine(driver)
+  const engine = initQueryEngine(conn)
 
   console.log('[nodejs] connecting...')
   await engine.connect('trace')
   console.log('[nodejs] connected')
 
-  console.log('[nodejs] isHealthy', await driver.isHealthy())
+  console.log('[nodejs] version', await conn.version())
+  console.log('[nodejs] isHealthy', await conn.isHealthy())
 
   // Smoke test for Neon that ensures we're able to decode every common data type.
   const resultSet = await engine.query(`{
@@ -72,7 +73,7 @@ async function main() {
 
   // Close the database connection. This is required to prevent the process from hanging.
   console.log('[nodejs] closing database connection...')
-  await driver.close()
+  await conn.close()
   console.log('[nodejs] closed database connection')
 
   process.exit(0)
